refactor(user): 네이밍 변경, 패키지 분리
