--- src/main/java/com/project/bumawiki/domain/user/presentation/dto/UserAuthorityRequestDto.java
@@ -1,11 +1,0 @@
-package com.project.bumawiki.domain.user.presentation.dto;
-
-import com.project.bumawiki.domain.user.domain.authority.Authority;
-
-import jakarta.validation.constraints.NotNull;
-
-public record UserAuthorityRequestDto(
-	@NotNull Long id,
-	@NotNull Authority authority
-) {
-}
