+++ src/main/java/com/project/bumawiki/domain/user/presentation/dto/response/SimpleUserResponseDto.java
@@ -1,15 +1,15 @@
-package com.project.bumawiki.domain.user.presentation.dto;
+package com.project.bumawiki.domain.user.presentation.dto.response;
 
 import com.project.bumawiki.domain.user.domain.User;
 
-public record SimpleUserDto(
+public record SimpleUserResponseDto(
 	Long id,
 	String email,
 	String nickName,
 	String name
 ) {
 
-	public SimpleUserDto(User user) {
+	public SimpleUserResponseDto(User user) {
 		this(
 			user.getId(),
 			user.getEmail(),
